--- api/src/main/java/io/minio/credentials/AnonymousCredentialsProvider.java
@@ -1,26 +1,0 @@
-/*
- * MinIO Java SDK for Amazon S3 Compatible Cloud Storage, (C) 2020 MinIO, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.minio.credentials;
-
-import io.minio.messages.Credentials;
-
-public class AnonymousCredentialsProvider implements CredentialsProvider {
-
-  @Override
-  public Credentials fetch() {
-    return Credentials.EMPTY;
-  }
-}
