+++ api/src/main/java/io/minio/credentials/MinioEnvironmentProvider.java
@@ -17,23 +17,16 @@ package io.minio.credentials;
 
 import io.minio.messages.Credentials;
 import java.time.ZonedDateTime;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.List;
-import javax.annotation.Nonnull;
 
 @SuppressWarnings("unused")
-public class AwsEnvironmentCredentialsProvider extends EnvironmentCredentialsProvider {
+public class MinioEnvironmentProvider extends EnvironmentProvider {
 
-  private static final List<String> ACCESS_KEY_ALIASES =
-      Arrays.asList("AWS_ACCESS_KEY_ID", "AWS_ACCESS_KEY");
-  private static final List<String> SECRET_KEY_ALIASES =
-      Arrays.asList("AWS_SECRET_ACCESS_KEY", "AWS_SECRET_KEY");
-  private static final String SESSION_TOKEN_ALIAS = "AWS_SESSION_TOKEN";
+  private static final String ACCESS_KEY_ALIAS = "MINIO_ACCESS_KEY";
+  private static final String SECRET_KEY_ALIAS = "MINIO_SECRET_KEY";
 
   private Credentials credentials;
 
-  public AwsEnvironmentCredentialsProvider() {
+  public MinioEnvironmentProvider() {
     credentials = readCredentials();
   }
 
@@ -52,20 +45,10 @@ public class AwsEnvironmentCredentialsProvider extends EnvironmentCredentialsPro
   }
 
   private Credentials readCredentials() {
-    final String accessKey = readFirst(ACCESS_KEY_ALIASES);
-    final String secretKey = readFirst(SECRET_KEY_ALIASES);
+    final String accessKey = readProperty(ACCESS_KEY_ALIAS);
+    final String secretKey = readProperty(SECRET_KEY_ALIAS);
     final ZonedDateTime lifeTime = ZonedDateTime.now().plus(REFRESHED_AFTER);
-    final String sessionToken = readProperty(SESSION_TOKEN_ALIAS);
-    return new Credentials(accessKey, secretKey, lifeTime, sessionToken);
-  }
-
-  private String readFirst(@Nonnull Collection<String> propertyKeys) {
-    for (String propertyKey : propertyKeys) {
-      final String value = readProperty(propertyKey);
-      if (value != null) {
-        return value;
-      }
-    }
-    throw new IllegalStateException("Can't find env variables for " + propertyKeys);
+    //noinspection ConstantConditions
+    return new Credentials(accessKey, secretKey, lifeTime);
   }
 }
