+++ api/src/main/java/io/minio/credentials/WebIdentityProvider.java
@@ -18,9 +18,9 @@ package io.minio.credentials;
 import io.minio.Xml;
 import io.minio.errors.InvalidResponseException;
 import io.minio.errors.XmlParserException;
-import io.minio.messages.AssumeRoleWithClientGrantsResponse;
-import io.minio.messages.ClientGrantsToken;
+import io.minio.messages.AssumeRoleWithWebIdentityResponse;
 import io.minio.messages.Credentials;
+import io.minio.messages.WebIdentityToken;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
@@ -31,20 +31,20 @@ import okhttp3.Response;
 import okhttp3.ResponseBody;
 
 @SuppressWarnings("unused")
-public class ClientGrantsCredentialsProvider extends StsCredentialsProvider {
+public class WebIdentityProvider extends StsProvider {
 
   private Credentials credentials;
-  private final Supplier<ClientGrantsToken> tokenProducer;
+  private final Supplier<WebIdentityToken> tokenProducer;
 
-  public ClientGrantsCredentialsProvider(
-      @Nonnull String stsEndpoint, @Nonnull Supplier<ClientGrantsToken> tokenProducer) {
+  public WebIdentityProvider(
+      @Nonnull String stsEndpoint, @Nonnull Supplier<WebIdentityToken> tokenProducer) {
     super(stsEndpoint);
     this.tokenProducer = Objects.requireNonNull(tokenProducer, "Token producer must not be null");
   }
 
   /**
-   * Returns a pointer to a new, temporary credentials, obtained via STS assume role with client
-   * grants api.
+   * Returns a pointer to a new, temporary credentials, obtained via STS assume role with web
+   * identity api.
    *
    * @return temporary credentials to access minio api.
    */
@@ -62,7 +62,7 @@ public class ClientGrantsCredentialsProvider extends StsCredentialsProvider {
             throw new IllegalStateException("Received empty response");
           }
           credentials =
-              Xml.unmarshal(AssumeRoleWithClientGrantsResponse.class, body.charStream())
+              Xml.unmarshal(AssumeRoleWithWebIdentityResponse.class, body.charStream())
                   .credentials();
         } catch (XmlParserException | IOException | InvalidResponseException e) {
           throw new IllegalStateException("Failed to process STS call", e);
@@ -74,11 +74,11 @@ public class ClientGrantsCredentialsProvider extends StsCredentialsProvider {
 
   @Override
   protected Map<String, String> queryParams() {
-    final ClientGrantsToken grantsToken = tokenProducer.get();
+    final WebIdentityToken grantsToken = tokenProducer.get();
     final Map<String, String> queryParamenters = new HashMap<>();
-    queryParamenters.put("Action", "AssumeRoleWithClientGrants");
+    queryParamenters.put("Action", "AssumeRoleWithWebIdentity");
     queryParamenters.put("DurationSeconds", tokenDuration(grantsToken.expiredAfter()));
-    queryParamenters.put("Token", grantsToken.token());
+    queryParamenters.put("WebIdentityToken", grantsToken.token());
     queryParamenters.put("Version", "2011-06-15");
     if (grantsToken.policy() != null) {
       queryParamenters.put("Policy", grantsToken.policy());
