+++ api/src/main/java/io/minio/messages/AssumeRoleWithClientGrantsResponse.java
@@ -17,16 +17,18 @@ package io.minio.messages;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Namespace;
+import org.simpleframework.xml.Path;
 import org.simpleframework.xml.Root;
 
 @Root(name = "AssumeRoleWithClientGrantsResponse", strict = false)
 @Namespace(reference = "https://sts.amazonaws.com/doc/2011-06-15/")
 public class AssumeRoleWithClientGrantsResponse {
 
-  @Element(name = "AssumeRoleWithClientGrantsResult")
-  private AssumeRoleWithClientGrantsResult clientGrantsResult;
+  @Path(value = "AssumeRoleWithClientGrantsResult")
+  @Element(name = "Credentials")
+  private Credentials credentials;
 
   public Credentials credentials() {
-    return clientGrantsResult == null ? null : clientGrantsResult.credentials();
+    return credentials;
   }
 }
